\name{rho.biweight}
\alias{rho.biweight}
\title{Tukey's biweight rho function}
\description{
Computes Tukey's biweight rho function with constant c for all 
values in the vector x.
}
\usage{
rho.biweight(x, c)
}
\arguments{
  \item{x}{a vector.}
  \item{c}{a scalar standing for the constant in the Tukey's biweight rho function.}
}
\seealso{psi.bisquare}
\keyword{robust}
\keyword{multivariate}
